+++ EnginDatabase/BackOffice/TypeCheckList.sql
@@ -1,11 +1,13 @@
 ﻿CREATE TABLE [dbo].[TypeCheckList] (
-    [Id]        INT            IDENTITY (1, 1) NOT NULL,
+    [Id]        BIGINT         IDENTITY (1, 1) NOT NULL,
     [Name]      NVARCHAR (MAX) NOT NULL,
     [CreatedOn] DATETIME       NOT NULL,
     [CreatedBy] NVARCHAR (128) NOT NULL,
     CONSTRAINT [PK_dbo.TypeCheckList] PRIMARY KEY CLUSTERED ([Id] ASC),
     CONSTRAINT [FK_dbo.TypeCheckList_dbo.AspNetUsers_CreatedBy] FOREIGN KEY ([CreatedBy]) REFERENCES [dbo].[AspNetUsers] ([Id])
 );
+
+
