+++ EnginDatabase/dbo/Tables/TypeEngin.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[TypeEngin] (
     [Id]              BIGINT         IDENTITY (1, 1) NOT NULL,
-    [TypeCheckListId] INT            NOT NULL,
+    [TypeCheckListId] BIGINT         NOT NULL,
     [Name]            NVARCHAR (MAX) NOT NULL,
     [DureeEstimative] NVARCHAR (MAX) NULL,
     [CreatedOn]       DATETIME       NOT NULL,
@@ -9,6 +9,8 @@
     CONSTRAINT [FK_dbo.TypeEngin_dbo.AspNetUsers_CreatedBy] FOREIGN KEY ([CreatedBy]) REFERENCES [dbo].[AspNetUsers] ([Id]),
     CONSTRAINT [FK_dbo.TypeEngin_dbo.TypeCheckList_TypeCheckListId] FOREIGN KEY ([TypeCheckListId]) REFERENCES [dbo].[TypeCheckList] ([Id])
 );
+
+
 
 
 GO
